refactor(library/Attempt): extracts "Outcome/Success.ts" from "Outcome/index.ts"
